images/loader/requirements.txt
clickhouse-driver>=0.2.7,<1
google-cloud-storage>=2.10,<3
google-cloud-secret-manager>=2.16,<3
google-cloud-firestore>=2.13,<3
